+++ src/crimson/os/CMakeLists.txt
@@ -1,8 +1,7 @@
 add_library(crimson-os
-  cyan_store.cc
-  cyan_collection.cc
-  cyan_object.cc
   futurized_store.cc
   ${PROJECT_SOURCE_DIR}/src/os/Transaction.cc)
+add_subdirectory(cyanstore)
 target_link_libraries(crimson-os
+  crimson-cyanstore
   crimson)
