+++ src/crimson/os/futurized_store.cc
@@ -1,5 +1,5 @@
 #include "futurized_store.h"
-#include "cyan_store.h"
+#include "cyanstore/cyan_store.h"
 
 namespace crimson::os {
